requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
pillow
pandas
requests
